Additions [#METR-21504].
